+++ packages/components/atoms/f-button/test-utils/component-objects/f-button.component.js
@@ -1,5 +1,17 @@
-const buttonComponent = () => $('[data-test-id="action-button-component"]');
+const Page = require('@justeat/f-wdio-utils/src/page.object');
 
-exports.waitForButtonComponent = () => buttonComponent().waitForExist();
+module.exports = class Button extends Page {
+    get component () { return $('[data-test-id="action-button-component"]'); }
 
-exports.isButtonComponentDisplayed = () => buttonComponent().isDisplayed();
+    open () {
+        super.openComponent('atom', 'button-component');
+    }
+
+    waitForComponent () {
+        this.component.waitForExist();
+    }
+
+    isComponentDisplayed () {
+        return this.component.isDisplayed();
+    }
+};