+++ packages/components/organisms/f-checkout/src/components/_tests/Checkout.test.js
@@ -8,7 +8,7 @@ import VueCheckout from '../Checkout.vue';
 import EventNames from '../../event-names';
 
 import {
-    defaultState, defaultActions, i18n, createStore, $logger
+    defaultCheckoutState, defaultCheckoutActions, i18n, createStore, $logger
 } from './helpers/setup';
 
 const localVue = createLocalVue();
@@ -147,7 +147,7 @@ describe('Checkout', () => {
             it('should display the address block if set to `delivery`', async () => {
                 // Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_DELIVERY }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_DELIVERY }),
                     i18n,
                     localVue,
                     propsData
@@ -162,7 +162,7 @@ describe('Checkout', () => {
             it('should not display the address block if set to `collection`', async () => {
                 // Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                     i18n,
                     localVue,
                     propsData
@@ -259,7 +259,7 @@ describe('Checkout', () => {
 
                 // Act
                 shallowMount(VueCheckout, {
-                    store: createStore(defaultState, { ...defaultActions, setAuthToken }),
+                    store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, setAuthToken }),
                     i18n,
                     localVue,
                     propsData
@@ -316,7 +316,7 @@ describe('Checkout', () => {
             it('should return `true` if `serviceType` is set to Delivery', () => {
                 // Arrange and Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_DELIVERY }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_DELIVERY }),
                     i18n,
                     localVue,
                     propsData
@@ -329,7 +329,7 @@ describe('Checkout', () => {
             it('should return `false` if `serviceType` is set to Collection', () => {
                 // Arrange and Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                     i18n,
                     localVue,
                     propsData
@@ -372,7 +372,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: true,
                         serviceType: CHECKOUT_METHOD_DELIVERY,
                         address: {}
@@ -393,7 +393,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: false,
                         serviceType: CHECKOUT_METHOD_DELIVERY,
                         address: {}
@@ -414,7 +414,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: true,
                         serviceType: CHECKOUT_METHOD_COLLECTION,
                         address: {}
@@ -433,7 +433,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: true,
                         serviceType: CHECKOUT_METHOD_DELIVERY,
                         address: { line1: 'Fleet Place House', postcode: 'EC4M 7RF', city: 'London' }
@@ -453,14 +453,19 @@ describe('Checkout', () => {
     });
 
     describe('mounted ::', () => {
+        let initialiseSpy;
+        let trackInitialLoadSpy;
+
+        beforeEach(() => {
+            initialiseSpy = jest.spyOn(VueCheckout.methods, 'initialise');
+            trackInitialLoadSpy = jest.spyOn(VueCheckout.methods, 'trackInitialLoad');
+        });
+
         afterEach(() => {
             jest.clearAllMocks();
         });
 
         it('should call `initialise`', () => {
-            // Arrange
-            const initialiseSpy = jest.spyOn(VueCheckout.methods, 'initialise');
-
             // Act
             shallowMount(VueCheckout, {
                 store: createStore(),
@@ -471,6 +476,21 @@ describe('Checkout', () => {
 
             // Assert
             expect(initialiseSpy).toHaveBeenCalled();
+        });
+
+        it('should call `trackInitialLoad`', async () => {
+            // Act
+            const wrapper = shallowMount(VueCheckout, {
+                store: createStore(),
+                i18n,
+                localVue,
+                propsData
+            });
+
+            await wrapper.vm.initialise();
+
+            // Assert
+            expect(trackInitialLoadSpy).toHaveBeenCalled();
         });
     });
 
@@ -628,7 +648,7 @@ describe('Checkout', () => {
                     const loadCheckoutSpy = jest.spyOn(VueCheckout.methods, 'loadCheckout');
 
                     shallowMount(VueCheckout, {
-                        store: createStore({ ...defaultState, isLoggedIn: true }),
+                        store: createStore({ ...defaultCheckoutState, isLoggedIn: true }),
                         i18n,
                         localVue,
                         propsData
@@ -748,7 +768,7 @@ describe('Checkout', () => {
                 describe('when all the fields are populated correctly', () => {
                     beforeEach(() => {
                         wrapper = mount(VueCheckout, {
-                            store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                            store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                             i18n,
                             localVue,
                             propsData,
@@ -793,10 +813,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: ''
                             }
                         }),
@@ -822,10 +842,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: '077777'
                             }
                         }),
@@ -852,10 +872,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: 'hs;-j`$e&1l'
                             }
                         }),
@@ -881,7 +901,7 @@ describe('Checkout', () => {
                 it('should not create validations for address', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
-                        store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                        store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                         i18n,
                         localVue,
                         propsData,
@@ -898,10 +918,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: ''
                             }
                         }),
@@ -927,7 +947,7 @@ describe('Checkout', () => {
                 describe('when all the fields are populated correctly', () => {
                     beforeEach(() => {
                         wrapper = mount(VueCheckout, {
-                            store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_DELIVERY }),
+                            store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_DELIVERY }),
                             i18n,
                             localVue,
                             propsData,
@@ -972,10 +992,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 line1: ''
                             }
                         }),
@@ -1001,10 +1021,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 city: ''
                             }
                         }),
@@ -1030,10 +1050,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: ''
                             }
                         }),
@@ -1059,10 +1079,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: '?!hdb-se'
                             }
                         }),
@@ -1089,10 +1109,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: 'EC4M 7R'
                             }
                         }),
@@ -1125,10 +1145,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: 'EC4M 7R'
                             }
                         }),
@@ -1158,7 +1178,7 @@ describe('Checkout', () => {
                     const setupGuestUserSpy = jest.spyOn(VueCheckout.methods, 'setupGuestUser');
                     const wrapper = shallowMount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             isLoggedIn: false
                         }),
                         i18n,
@@ -1180,7 +1200,7 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             isLoggedIn: false
                         }),
@@ -1204,10 +1224,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 firstName: ''
                             },
                             isLoggedIn: false
@@ -1234,10 +1254,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 lastName: ''
                             },
                             isLoggedIn: false
@@ -1264,10 +1284,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 email: ''
                             },
                             isLoggedIn: false
@@ -1294,10 +1314,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 email: '£Gs7asd263('
                             },
                             isLoggedIn: false
@@ -1327,7 +1347,7 @@ describe('Checkout', () => {
                     const setupGuestUserSpy = jest.spyOn(VueCheckout.methods, 'setupGuestUser');
                     const wrapper = shallowMount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             authToken: 'sampleToken',
                             isLoggedIn: true
                         }),
@@ -1350,7 +1370,7 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = shallowMount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             authToken: 'sampleToken',
                             isLoggedIn: true
                         }),
@@ -1433,7 +1453,7 @@ describe('Checkout', () => {
                 const createGuestUserSpy = jest.spyOn(VueCheckout.methods, 'createGuestUser');
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         customer
                     }),
                     i18n,
@@ -1462,7 +1482,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, createGuestUser: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, createGuestUser: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1531,7 +1551,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getCheckout: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getCheckout: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1597,7 +1617,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getAvailableFulfilment: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getAvailableFulfilment: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1663,7 +1683,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getBasket: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getBasket: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1726,7 +1746,7 @@ describe('Checkout', () => {
                 it('should emit failure event and set `hasCheckoutLoadedSuccessfully` to `false`', async () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getAddress: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getAddress: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData
@@ -1922,18 +1942,48 @@ describe('Checkout', () => {
         describe('onFormSubmit ::', () => {
             let isFormValidSpy;
             let submitCheckoutSpy;
+            let trackFormInteractionSpy;
 
             beforeEach(() => {
                 isFormValidSpy = jest.spyOn(VueCheckout.methods, 'isFormValid');
                 submitCheckoutSpy = jest.spyOn(VueCheckout.methods, 'submitCheckout');
+                trackFormInteractionSpy = jest.spyOn(VueCheckout.methods, 'trackFormInteraction');
             });
 
             afterEach(() => {
                 jest.clearAllMocks();
             });
 
-            it('should emit `CheckoutValidationError` with validation state if form is invalid', async () => {
+            it('should call `trackFormInteractionSpy` with correct `action`', async () => {
                 // Arrange
+                isFormValidSpy.mockReturnValue(true);
+
+                const wrapper = mount(VueCheckout, {
+                    store: createStore(),
+                    i18n,
+                    localVue,
+                    propsData,
+                    mocks: {
+                        $v,
+                        $logger
+                    }
+                });
+
+                const payload = {
+                    action: 'submit'
+                };
+
+                // Act
+                await wrapper.vm.onFormSubmit();
+
+                // Assert
+                expect(trackFormInteractionSpy).toHaveBeenCalledWith(payload);
+            });
+
+            describe('when the form is Invalid', () => {
+                let wrapper;
+                let getFormValidationStateSpy;
+
                 const mockValidationState = {
                     validFields: [
                         'customer.mobileNumber',
@@ -1944,76 +1994,117 @@ describe('Checkout', () => {
                     invalidFields: []
                 };
 
-                const getFormValidationStateSpy = jest.spyOn(validations, 'getFormValidationState');
+                beforeEach(() => {
+                    isFormValidSpy.mockReturnValue(false);
 
-                getFormValidationStateSpy.mockReturnValue(mockValidationState);
-                isFormValidSpy.mockReturnValue(false);
+                    getFormValidationStateSpy = jest.spyOn(validations, 'getFormValidationState');
+                    getFormValidationStateSpy.mockReturnValue(mockValidationState);
 
-                const wrapper = mount(VueCheckout, {
-                    store: createStore(),
-                    i18n,
-                    localVue,
-                    propsData,
-                    mocks: {
-                        $v,
-                        $logger
-                    }
+                    wrapper = mount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $v,
+                            $logger
+                        }
+                    });
                 });
 
-                // Act
-                await wrapper.vm.onFormSubmit();
+                it('should emit `CheckoutValidationError`', async () => {
+                    // Act
+                    await wrapper.vm.onFormSubmit();
 
-                // Assert
-                expect(wrapper.emitted(EventNames.CheckoutValidationError).length).toBe(1);
-                expect(wrapper.emitted(EventNames.CheckoutValidationError)[0][0]).toEqual(mockValidationState);
-            });
-
-            it('should try to call `submitCheckout` if form is Valid', async () => {
-                // Arrange
-                isFormValidSpy.mockReturnValue(true);
-
-                const wrapper = mount(VueCheckout, {
-                    store: createStore(),
-                    i18n,
-                    localVue,
-                    propsData,
-                    mocks: {
-                        $v,
-                        $logger
-                    }
+                    // Assert
+                    expect(wrapper.emitted(EventNames.CheckoutValidationError).length).toBe(1);
                 });
 
-                // Act
-                await wrapper.vm.onFormSubmit();
+                it('should call `trackFormInteraction` with an action `inline_error` and validation errors', async () => {
+                    // Arrange
+                    const payload = {
+                        action: 'inline_error',
+                        error: mockValidationState.invalidFields
+                    };
 
-                // Assert
-                expect(submitCheckoutSpy).toHaveBeenCalled();
+                    // Act
+                    await wrapper.vm.onFormSubmit();
+
+                    // Assert
+                    expect(trackFormInteractionSpy).toHaveBeenCalledWith(payload);
+                });
             });
 
-            it('should call `handleErrorState` if `submitCheckout` returns an error', async () => {
-                // Arrange
-                const handleErrorStateSpy = jest.spyOn(VueCheckout.methods, 'handleErrorState');
+            describe('when the form is valid', () => {
+                let wrapper;
+
+                beforeEach(() => {
+                    isFormValidSpy.mockReturnValue(true);
+
+                    wrapper = mount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $v,
+                            $logger
+                        }
+                    });
+                });
+
+                it('should try to call `submitCheckout`', async () => {
+                    // Act
+                    await wrapper.vm.onFormSubmit();
+
+                    // Assert
+                    expect(submitCheckoutSpy).toHaveBeenCalled();
+                });
+
+                it('should call `trackFormInteractionSpy` with correct `action`', async () => {
+                    // Arrange
+                    const payload = {
+                        action: 'success'
+                    };
+
+                    // Act
+                    await wrapper.vm.onFormSubmit();
+
+                    // Assert
+                    expect(trackFormInteractionSpy).toHaveBeenCalledWith(payload);
+                });
+            });
+
+            describe('when `submitCheckout` returns an error', () => {
+                let wrapper;
+                let handleErrorStateSpy;
                 const error = new Error('errorMessage');
 
-                submitCheckoutSpy.mockImplementation(() => { throw error; });
-                isFormValidSpy.mockReturnValue(true);
+                beforeEach(() => {
+                    handleErrorStateSpy = jest.spyOn(VueCheckout.methods, 'handleErrorState');
 
-                const wrapper = mount(VueCheckout, {
-                    store: createStore(),
-                    i18n,
-                    localVue,
-                    propsData,
-                    mocks: {
-                        $v,
-                        $logger
-                    }
+                    submitCheckoutSpy.mockImplementation(() => { throw error; });
+                    isFormValidSpy.mockReturnValue(true);
+
+                    wrapper = mount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $v,
+                            $logger
+                        }
+                    });
                 });
 
-                // Act
-                await wrapper.vm.onFormSubmit();
+                it('should call `handleErrorState`', async () => {
+                    // Act
+                    await wrapper.vm.onFormSubmit();
 
-                // Assert
-                expect(handleErrorStateSpy).toHaveBeenCalledWith(error);
+                    // Assert
+                    expect(handleErrorStateSpy).toHaveBeenCalledWith(error);
+                });
             });
 
             it('should call `logWarn` if form is Invalid', async () => {
@@ -2059,7 +2150,7 @@ describe('Checkout', () => {
                 wrapper.vm.isValidPhoneNumber();
 
                 // Assert
-                expect(isValidPhoneNumberSpy).toHaveBeenCalledWith(defaultState.customer.mobileNumber, i18n.locale);
+                expect(isValidPhoneNumberSpy).toHaveBeenCalledWith(defaultCheckoutState.customer.mobileNumber, i18n.locale);
             });
         });
 
@@ -2083,7 +2174,7 @@ describe('Checkout', () => {
                 wrapper.vm.isValidPostcode();
 
                 // Assert
-                expect(isValidPostcodeSpy).toHaveBeenCalledWith(defaultState.address.postcode, i18n.locale);
+                expect(isValidPostcodeSpy).toHaveBeenCalledWith(defaultCheckoutState.address.postcode, i18n.locale);
             });
         });
 
@@ -2134,7 +2225,7 @@ describe('Checkout', () => {
                 jest.advanceTimersByTime(1000);
 
                 // Assert
-                expect(windowLocationSpy).toHaveBeenCalledWith(`${paymentPageUrlPrefix}/${defaultState.orderId}`);
+                expect(windowLocationSpy).toHaveBeenCalledWith(`${paymentPageUrlPrefix}/${defaultCheckoutState.orderId}`);
             });
 
             it('should not redirect to the payment page before 1 second', () => {
@@ -2176,7 +2267,7 @@ describe('Checkout', () => {
                         basketId,
                         applicationId: 7,
                         customerNotes: {
-                            noteForRestaurant: defaultState.userNote
+                            noteForRestaurant: defaultCheckoutState.userNote
                         },
                         applicationName,
                         applicationVersion: '1',
@@ -2189,7 +2280,7 @@ describe('Checkout', () => {
 
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         basket: {
                             id: basketId
                         }